+++ packages/terra-dropdown-button/src/SplitButton.jsx
@@ -60,6 +60,11 @@ const propTypes = {
    */
   // eslint-disable-next-line react/forbid-prop-types
   metaData: PropTypes.object,
+  /**
+   * Sets the custom properties for button.
+   */
+  // eslint-disable-next-line react/forbid-prop-types
+  buttonAttrs: PropTypes.object,
 };
 
 const defaultProps = {
@@ -67,6 +72,7 @@ const defaultProps = {
   isCompact: false,
   isDisabled: false,
   variant: 'neutral',
+  buttonAttrs: {},
 };
 
 class SplitButton extends React.Component {
@@ -191,6 +197,7 @@ class SplitButton extends React.Component {
       intl,
       requestClose,
       metaData,
+      buttonAttrs,
       ...customProps
     } = this.props;
 
@@ -226,6 +233,15 @@ class SplitButton extends React.Component {
       theme.className,
     );
 
+    let buttonAriaLabel = '';
+    const modifiedButtonAttrs = { ...buttonAttrs };
+    if (modifiedButtonAttrs && modifiedButtonAttrs['aria-label']) {
+      buttonAriaLabel = modifiedButtonAttrs['aria-label'];
+      delete modifiedButtonAttrs['aria-label'];
+    }
+    const customLabel = (selectText) ? `${selectText}, ${selectedLabel}, ${caretLabel}` : caretLabel;
+    buttonAriaLabel = `${customLabel}${buttonAriaLabel ? `, ${buttonAriaLabel}` : ''}`;
+
     return (
       <DropdownButtonBase
         {...customProps}
@@ -252,6 +268,7 @@ class SplitButton extends React.Component {
           {primaryOptionLabel}
         </button>
         <button
+          {...modifiedButtonAttrs}
           type="button"
           onClick={this.handleDropdownButtonClick}
           onKeyDown={this.handleCaretKeyDown}
@@ -262,7 +279,7 @@ class SplitButton extends React.Component {
           aria-disabled={isDisabled}
           aria-expanded={isOpen}
           aria-haspopup="menu"
-          aria-label={selectText ? `${selectText}, ${selectedLabel}, ${caretLabel}` : caretLabel}
+          aria-label={buttonAriaLabel}
           onFocus={this.handleFocus}
           onBlur={this.handleBlur}
           ref={this.setButtonNode}